# docs/swagger.yaml
openapi: 3.0.0
info:
  title: API для расчета лекарственных препаратов для детей
  description: REST API для работы с лекарствами, заявками и пользователями
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: Локальный сервер разработки
paths:
...
  /prescriptions/{id}/complete:
    put:
      summary: Завершить заявку модератором
      tags:
        - Заявки (Prescription)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Заявка завершена
  /prescriptions/{id}/reject:
    put:
      summary: Отклонить заявку модератором
      tags:
        - Заявки (Prescription)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Заявка отклонена
...